Scheduling: Implemented the interrupt handler for the PIT Chip
Implemented the interrupt handler for the PIT Chip.

+++ kernel/scheduling/pit/pit.cpp
@@ -1,5 +1,5 @@
 #include "pit.h"
-#include "../IO.h"
+#include "../../IO.h"
 
 namespace PIT{
     double TimeSinceBoot = 0;
@@ -13,8 +13,8 @@ namespace PIT{
         }
     }
 
-    void sleep(double milliseconds){
-        sleepd(milliseconds/1000);
+    void sleep(uint64_t milliseconds){
+        sleepd((double)milliseconds/1000);
     }
 
     void setDivisor(uint16_t divisor){
