+++ kernel/scheduling/pit/pit.h
@@ -6,7 +6,7 @@ namespace PIT{
     const uint64_t BaseFrequency = 1193182;
 
     void sleepd(double seconds);
-    void sleep(double milliSeconds);
+    void sleep(uint64_t milliSeconds);
 
     void setDivisor(uint16_t divisor);
     uint64_t getFrequency();
